requirements.txt
pandas==2.2.2
numpy==2.0.0
Telethon==1.36.0
google-cloud-bigquery==3.25.0
Flask==3.0.3